refactor: split mini app god-component
App.tsx (1033 lines) becomes a 198-line orchestration layer: pure
helpers move to appModel.ts, the server-sync lifecycle to the
useMiniAppSnapshots hook, presentational sections to components/.
Code moved verbatim; guardrail test anchors relocated with their code.

findings)
